refactor plotting
