+++ oceanarray/analysis/_clock_diagnostics.py
@@ -19,7 +19,7 @@ import pandas as pd
 import xarray as xr
 import yaml
 
-from oceanarray.analysis._science import lag_correlation
+from oceanarray.analysis.tseries import lag_correlation
 from oceanarray.legacy import find_deployment
